refactor(compile): extract compiler guards
