+++ jpa-rsql-search-core/src/test/java/io/github/ggomarighetti/jparsqlsearch/compile/SearchPageablePropertyTest.java
@@ -1,6 +1,5 @@
 package io.github.ggomarighetti.jparsqlsearch.compile;
 
-import io.github.ggomarighetti.jparsqlsearch.compile.SearchCompiler.SearchPageableGuard;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
